Aktualisiere main.blocks, main.ts

+++ main.ts
@@ -1,12 +1,13 @@
 input.onButtonEvent(Button.A, input.buttonEventClick(), function () {
     Player = game.createSprite(2, 4)
     while (A == false) {
-        while (input.buttonIsPressed(Button.A)) {
+        if (input.buttonIsPressed(Button.A)) {
             Player.move(-1)
             basic.pause(100)
-        }
-        while (input.buttonIsPressed(Button.B)) {
+        } else if (input.buttonIsPressed(Button.B)) {
             Player.move(1)
+            basic.pause(100)
+        } else {
             basic.pause(100)
         }
     }
@@ -14,23 +15,29 @@ input.onButtonEvent(Button.A, input.buttonEventClick(), function () {
 let Player: game.LedSprite = null
 let Star: game.LedSprite = null
 let A = false
-while (A == false) {
-    basic.pause(100)
+let B = 1000
+if (A == false) {
+    basic.pause(B)
     Star = game.createSprite(randint(0, 4), 0)
-    basic.pause(100)
+    basic.pause(B)
     Star.turn(Direction.Right, 90)
     Star.move(1)
-    basic.pause(100)
+    basic.pause(B)
     Star.move(1)
-    basic.pause(100)
+    basic.pause(B)
     Star.move(1)
-    basic.pause(100)
+    basic.pause(B)
     Star.move(1)
-    basic.pause(100)
+    basic.pause(B)
     Star.delete()
+    B += -10
+} else if (Star.isTouching(Player)) {
+    A = true
+} else {
+	
 }
 basic.forever(function () {
-    if (Star.isTouching(Player)) {
-        A = true
+    if (true) {
+        basic.setLedColor(0xff0000)
     }
 })
